requirements.txt
Django==2.1.3
attrs==18.2.0
pytz==2018.7
